requirements.txt
fastapi
uvicorn
transformers==4.41.0
torch
python-dotenv
accelerate
protobuf
python-json-logger